+++ apiserver/testing/fakeapi.go
@@ -21,10 +21,10 @@ import (
 	"github.com/juju/juju/core/flightrecorder"
 	"github.com/juju/juju/core/trace"
 	jujuhttp "github.com/juju/juju/internal/http"
-	"github.com/juju/juju/internal/rpcreflect"
 	"github.com/juju/juju/internal/testing"
 	"github.com/juju/juju/rpc"
 	"github.com/juju/juju/rpc/jsoncodec"
+	"github.com/juju/juju/rpc/rpcreflect"
 )
 
 // Server represents a fake API server. It must be closed
