+++ generate/schemagen/gen/generator_test.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/juju/juju/apiserver/facade"
 	jsonschema "github.com/juju/juju/generate/schemagen/jsonschema-gen"
-	"github.com/juju/juju/internal/rpcreflect"
 	"github.com/juju/juju/internal/testhelpers"
+	"github.com/juju/juju/rpc/rpcreflect"
 )
 
 type GenSuite struct {
