+++ generate/schemagen/jsonschema-gen/reflect.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/juju/juju/internal/rpcreflect"
+	"github.com/juju/juju/rpc/rpcreflect"
 )
 
 // Cache the following types to special case them for speed.
